requirements.txt
# Core dependencies
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0

# PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data processing
pandas==2.1.4
numpy==1.26.3

# API and web scraping (if needed)
requests==2.31.0
beautifulsoup4==4.12.3

# Utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.1

# Development
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0